# Conflicts:
#	src/main/java/com/woniu/hospital_information_system/controller/EmployeeController.java
#	src/main/java/com/woniu/hospital_information_system/service/EmployeeService.java
#	src/main/java/com/woniu/hospital_information_system/service/impl/EmployeeServiceImpl.java

+++ src/main/java/com/woniu/hospital_information_system/entity/VisitorInfo.java
@@ -21,7 +21,7 @@ public class VisitorInfo {
     private Integer gender;
     private Integer age;
     private String idNumber;
-    private Integer phone;
+    private BigInteger phone;
     private Integer unitId;
     private Integer doctorId;
     private Integer diseaseId;
